Added a single reference power spectrum as well as the mean of 100 PS in numpy files.
Edited map_power_beta to only calculate the bcls, not the beta map to improve speed

Edited the get_wigner_sum function to marginally improve speed. Optimization here is key as it is the dominant factor in computation time.

Ran some optimisations against reference powerspectra. So far no success. Either the optimizer doesn't move at all from starting amplitude, or else it moves, but returns terrible estimates, in particular it seems to like the value gamma~-15. It also seems to be favouring the fits at higher freqencies, as the fits at 30 GHz are consistently the worst of them all, and I'm unsure why that would be.