Added attempt to curve fit in 1d, not currently working for unknown reasons.
Attempting to optimize the get_wigner_sum function by vectorizing the double for loop. The idea is to create a (384,384,384) shape array storing the w3j (done) the 2*ell+1 factors in the 1,2 indices, the amp_cls in the 1 index, and the beta_cls in the 2 index. Then multiply these arrays together and np.sum over the ell1 and ell2 dimensions. However, this is not yet functioning, and the scaffolding work done is both slower and does not replicate same results as previous work.