Cleaned up the foreground_functions.py file to move functions that aren't being used currently to the end.
Attempted doing 2d fits with the full model. Minimize now runs in more reasonable time (up to ~80 sec). Making use of L-BFGS-B method makes all the parameters move from the initialising point. The minimization does a good job of extracting close to input alpha, beta, and gamma. However, it does a poor job of finding the correct amplitude. (It moves away from the input amplitude). This may be an issue with the way that the objective function is written. Two other observations:

 - cutting out the high ell (where the model and data don't match) doesn't help find the right amplitude

 - the fit seems to be prioritizing the higher amplitudes the most, at least by eye. The 30 GHz fit seems to always be the worst fit. This might be an indication of an error in the objective function.

To deal with the fact that the input model and the data do not match at high ell (an effect of the finite nside of the maps that make the data) we could try to fix this a few different ways:

 - we could make the data with much larger ell than we're actually trying to model and then just truncate

 - we could try to model this divergence somehow and adjust the objective functino to penalise the high ell for this. This modelling could be either empirical or we could try to work out the analytic effect of this finite effect.