Change objective function to run over ell=30 onwards only.
Plotted errors from finite nside on log log to confirm they were powerlaws

With appropriate contraints on A and gamma the optimizer can fine a good set of parameters with more freqencies (to help constrain beta). Going forward we can try fitting the logs of the data and model to standardise across a large dynamic range, and generating data with a higher nside than the model, then truncating to the appropriate length for fitting.