Major edits to foreground_functions.py. Rewrote everything to include dust terms.
Added new notebook to play with inclusion of dust foregrounds. Optimisation not quite working, unsure why, and I need to check the SED of dust in the functions file, as I currently just have to fudge the order of magnitude to get synch and dust to change places at some point.